requirements.txt
# MO AI Backend - Cloud Requirements

# Google Cloud Libraries
google-cloud-bigquery==3.14.0
google-cloud-aiplatform==1.38.0
google-cloud-storage==2.10.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.106.0

# AI/ML Libraries
google-generativeai==0.3.0
anthropic==0.7.10
openai==1.3.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# HTTP & Async
requests==2.31.0
httpx==0.25.0
aiohttp==3.9.0

# Data Processing
pydantic==2.4.2
python-dotenv==1.0.0
pytz==2023.3

# Utilities
click==8.1.7
tqdm==4.66.1
python-dateutil==2.8.2

# Logging & Monitoring
python-json-logger==2.0.7
structlog==23.2.0

# Testing (optional, remove for production)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development (optional, remove for production)
black==23.11.0
pylint==3.0.2
mypy==1.7.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.5.0

# Caching
redis==5.0.1

# Secrets Management
cryptography==41.0.7

# Performance
uvicorn==0.24.0
fastapi==0.104.1

# Environment
python-multipart==0.0.6